employee-api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
pydantic[email]==2.5.3
python-multipart==0.0.6
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
